requirements.txt
python-telegram-bot[all]>=20.0
asyncpg>=0.29.0
openai>=1.0.0
python-dotenv>=1.0.0
